Changing data structure for fitness scores. Fixing minor bugs.

+++ controllers/EpuckController/util/FilesFunctions.java
@@ -35,7 +35,7 @@ public class FilesFunctions {
 
     }
 
-    public static void logAllCompFit(BufferedWriter out, float [][] compFit, int generation){
+    public static void logAllCompFit(BufferedWriter out, double [][] compFit, int generation){
 
         try{
             out.write(""+generation);
@@ -73,7 +73,7 @@ public class FilesFunctions {
         }
     }
 
-    public static void logAllFitnesses(BufferedWriter out2, int generation, float[] gameFitness) {
+    public static void logAllFitnesses(BufferedWriter out2, int generation, double[] gameFitness) {
         try{
             out2.write(""+generation+", ");
             for(int i=0; i<gameFitness.length; i++) out2.write(""+gameFitness[i]+", ");
